+++ package.json
@@ -18,6 +18,7 @@
     "crypto-js": "^4.1.1",
     "date-fns": "^2.30.0",
     "firebase": "^9.22.0",
+    "jest": "^29.6.1",
     "js-cookie": "^3.0.5",
     "react": "^18.2.0",
     "react-cookie": "^4.1.1",
@@ -34,7 +35,7 @@
   "scripts": {
     "start": "react-scripts start",
     "build": "react-scripts build",
-    "test": "react-scripts test --verbose",
+    "test": "react-scripts test --verbose ",
     "eject": "react-scripts eject",
     "dev": "webpack-dev-server --mode development",
     "webpack": "webpack --watch"
