+++ src/components/createEvent/daytime/daytime.tsx
@@ -1,5 +1,7 @@
 import { Dispatch, SetStateAction, useState } from "react";
-import InputDaytime from "./inputDaytime";
+import styles from "../../../styles/event/create.module.css";
+import HandleBlur from "./handleBlur";
+import HandleChange from "./handleChange";
 
 export default function Daytime({
   startDate,
@@ -12,14 +14,45 @@ export default function Daytime({
   endDate: string;
   setEndDate: Dispatch<SetStateAction<string>>;
 }) {
+  const [error, setError] = useState("");
+
   return (
     <>
-      <InputDaytime
-        startDate={startDate}
-        endDate={endDate}
-        setStartDate={setStartDate}
-        setEndDate={setEndDate}
+      <input
+        type="date"
+        className={`${styles.date} ${error ? styles.errorInput : ""}`}
+        maxLength={300}
+        value={startDate}
+        id="startDate"
+        onChange={HandleChange({
+          setStartDate,
+          setEndDate,
+          error,
+          setError,
+          type: "start",
+        })}
       />
+      &nbsp; ～ &nbsp;
+      <input
+        type="date"
+        className={`${styles.date} ${error ? styles.errorInput : ""}`}
+        maxLength={300}
+        value={endDate}
+        id="endDate"
+        onChange={HandleChange({
+          setStartDate,
+          setEndDate,
+          error,
+          setError,
+          type: "end",
+        })}
+      />
+      {error && (
+        <div>
+          <span className={styles.span}>{error}</span>
+        </div>
+      )}
+      <HandleBlur startDate={startDate} endDate={endDate} setError={setError} />
     </>
   );
 }
