+++ src/components/createEvent/detail/detail.tsx
@@ -1,5 +1,9 @@
+// JSX表示のみコンポーネント
+
 import { Dispatch, SetStateAction, useState } from "react";
-import InputDetail from "./inputDetail";
+import styles from "../../../styles/event/create.module.css";
+import HandleDetailBlur from "../../createIsland/detail/handleBlur";
+import HandleDetailChange from "../../createIsland/detail/handleDetailChange";
 
 export default function EventDetail({
   detail,
@@ -8,9 +12,22 @@ export default function EventDetail({
   detail: string;
   setDetail: Dispatch<SetStateAction<string>>;
 }) {
+  const [error, setError] = useState("");
+
   return (
     <>
-      <InputDetail detail={detail} setDetail={setDetail} />
+      <textarea
+        className={`${styles.detail} ${error ? styles.errorInput : ""} `}
+        maxLength={250}
+        value={detail}
+        onChange={HandleDetailChange({ setDetail, error, setError })}
+      />
+      {error && (
+        <div>
+          <span className={styles.span}>{error}</span>
+        </div>
+      )}
+      <HandleDetailBlur detail={detail} setError={setError} type={"event"} />
     </>
   );
 }
