+++ src/components/createIsland/detail/detail.tsx
@@ -1,5 +1,9 @@
+// JSX表示のみコンポーネント
+
 import { Dispatch, SetStateAction, useState } from "react";
-import InputDetail from "./inputDetail";
+import styles from "../../../styles/island/createIsland.module.css";
+import HandleDetailChange from "./handleDetailChange";
+import HandleNameBlur from "./handleBlur";
 
 export default function Detail({
   detail,
@@ -8,9 +12,26 @@ export default function Detail({
   detail: string;
   setDetail: Dispatch<SetStateAction<string>>;
 }) {
+  const [error, setError] = useState("");
+
   return (
     <>
-      <InputDetail detail={detail} setDetail={setDetail} />
+      <textarea
+        className={`${styles.detail} ${error ? styles.errorInput : ""} `}
+        maxLength={250}
+        value={detail}
+        onChange={HandleDetailChange({
+          error,
+          setDetail,
+          setError,
+        })}
+      />
+      {error && (
+        <div>
+          <span className={styles.span}>{error}</span>
+        </div>
+      )}
+      <HandleNameBlur detail={detail} setError={setError} type={"island"} />
     </>
   );
 }
