+++ src/components/createIsland/islandName/islandName.tsx
@@ -1,16 +1,54 @@
+// JSX表示のみコンポーネント
+
 import { Dispatch, SetStateAction, useState } from "react";
-import InputName from "./inputName";
+import styles from "../../../styles/island/createIsland.module.css";
+import HandleNameBlur from "./handleNameBlur";
+import HandleNameChange from "./handleNameChange";
 
 export default function IslandName({
   islandName,
-  setIslandName,
+  setName,
 }: {
   islandName: string;
-  setIslandName: Dispatch<SetStateAction<string>>;
+  setName: Dispatch<SetStateAction<string>>;
 }) {
+  const [NameError, setNameError] = useState("");
+  const [nameAlreadyError, setNameAlreadyError] = useState("");
+
   return (
     <>
-      <InputName islandName={islandName} setName={setIslandName} />
+      <input
+        type="text"
+        className={`${styles.islandName} ${NameError ? styles.errorInput : ""}`}
+        maxLength={100}
+        value={islandName}
+        onChange={HandleNameChange({
+          NameError,
+          setName,
+          setNameError,
+          setNameAlreadyError,
+          nameAlreadyError,
+        })}
+        onBlur={() => {
+          HandleNameBlur({
+            Name: islandName,
+            setNameError,
+            setNameAlreadyError,
+            type: "island",
+          });
+        }}
+      />
+      <span className={styles.islandNameText}>&nbsp;島</span>
+      {NameError && (
+        <div>
+          <span className={styles.span}>{NameError}</span>
+        </div>
+      )}
+      {nameAlreadyError && (
+        <div>
+          <span className={styles.span}>{nameAlreadyError}</span>
+        </div>
+      )}
     </>
   );
 }
