+++ src/tests/createIsland/detail/detail.test.js
@@ -1,17 +1,17 @@
 import { render, screen, fireEvent } from "@testing-library/react";
-import EventDetail from "../../components/createEvent/detail/detail";
+import Detail from "../../../components/createIsland/detail/detail";
 
 describe("Detail", () => {
-  test("テキストエリアが表示されること", () => {
-    render(<EventDetail detail="" setDetail={() => {}} />);
+  test("テキストエリアが表示されるかどうか", () => {
+    render(<Detail detail="" setDetail={() => {}} />);
 
     const textareaElement = screen.getByRole("textbox");
     expect(textareaElement).toBeInTheDocument();
   });
 
-  test("入力値が更新されたら、活動内容が更新されること", () => {
+  test("入力値が更新されたら、活動内容が更新されるかどうか", () => {
     const setDetail = jest.fn();
-    render(<EventDetail detail="" setDetail={setDetail} />);
+    render(<Detail detail="" setDetail={setDetail} />);
 
     const textareaElement = screen.getByRole("textbox");
     fireEvent.change(textareaElement, { target: { value: "New Detail" } });
@@ -19,19 +19,18 @@ describe("Detail", () => {
     expect(setDetail).toHaveBeenCalledWith("New Detail");
   });
 
-  test("入力が空の場合にエラーが表示されること", () => {
-    render(<EventDetail detail="" setDetail={() => {}} />);
+  test("入力が空の場合にエラーが表示されるかどうか", () => {
+    render(<Detail detail="" setDetail={() => {}} />);
 
     const textareaElement = screen.getByRole("textbox");
-    // テキストエリアからフォーカスが外れたとき
     fireEvent.blur(textareaElement);
 
-    const errorElement = screen.getByText("※詳細内容は入力必須項目です");
+    const errorElement = screen.getByText("※活動内容は入力必須項目です");
     expect(errorElement).toBeInTheDocument();
   });
 
-  test("入力されている場合にはエラーが表示されないこと", () => {
-    render(<EventDetail detail="Valid Detail" setDetail={() => {}} />);
+  test("入力が有効な場合にはエラーが表示されないこと", () => {
+    render(<Detail detail="Valid Detail" setDetail={() => {}} />);
 
     const textareaElement = screen.getByRole("textbox");
     fireEvent.blur(textareaElement);
